pyproject.toml

[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["Temirlan Tyulyubayev <temikbjj@gmail.com>"]
packages = [
    { include = "app" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
sqlalchemy = ">=2.0.41,<3.0.0"
alembic = ">=1.16.1,<2.0.0"
asyncpg = ">=0.30.0,<0.31.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
pydantic-settings = ">=2.10.1,<3.0.0"
greenlet = ">=3.2.3,<4.0.0"
aiogram = ">=3.21.0,<4.0.0"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
